feat : 레디스 key expired 로 1초간격마다 쿼리가 나가도록 수정하려고 했으나 실패

+++ src/main/java/com/board/post/service/PostService.java
@@ -91,7 +91,7 @@ public class PostService {
     }
 
     @Transactional
-    public void addViews(Long postId, int increment) {
+    public void addViews(Long postId, long increment) {
         Post post = postRepository.findById(postId).orElseThrow(EntityNotFoundException::new);
         post.incrementViews(increment);
     }
